refactor: simplify imports with index in Application
