+++ src/application/protocols/db/dynamic/authentication/add-account-repository.ts
@@ -1,4 +1,4 @@
-import { type AddAccount } from '@/domain/usecases/add-account'
+import { type AddAccount } from '@/domain/usecases/authentication'
 
 export interface AddAccountRepository {
   add: (input: AddAccountRepository.Input) => Promise<boolean>
