+++ src/application/protocols/db/dynamic/authentication/index.ts
@@ -1,0 +1,1 @@
+export * from './add-account-repository'
