+++ src/application/protocols/db/static/authentication/check-account-by-email-repository.ts
