+++ src/application/protocols/db/static/authentication/get-account-by-email-repository.ts
