+++ src/application/protocols/db/static/authentication/index.ts
@@ -1,0 +1,2 @@
+export * from './check-account-by-email-repository'
+export * from './get-account-by-email-repository'
