+++ src/application/usecases/authentication/db-add-account.ts
@@ -1,7 +1,7 @@
-import { type CheckAccountByEmailRepository } from '@/application/protocols/db/check-account-by-email-repository'
-import { type Hasher } from '@/application/protocols/cryptography/hasher'
-import { type AddAccountRepository } from '@/application/protocols/db/add-account-repository'
-import { type AddAccount } from '@/domain/usecases/add-account'
+import { type CheckAccountByEmailRepository } from '@/application/protocols/db/static/authentication'
+import { type Hasher } from '@/application/protocols/cryptography'
+import { type AddAccountRepository } from '@/application/protocols/db/dynamic/authentication'
+import { type AddAccount } from '@/domain/usecases/authentication'
 
 export class DbAddAccount implements AddAccount {
   constructor(
